+++ docs/.vitepress/ext/index.ts
@@ -1,3 +1,3 @@
-﻿export * as Locales from "./locales";
-export * from "./markdown";
-export * from "./vite";
+﻿export * as Locales from "./locales/locales";
+export * from "./markdown/markdown";
+export * from "./vite/vite";
