+++ docs/.vitepress/ext/language/naniscript-theme.json
@@ -116,7 +116,6 @@
                 "foreground": "#d14e4e"
             }
         },
-
         {
             "settings": {
                 "foreground": "#D4D4D4"
